Bubble chart fix
If the data source has 3 columns the third column is now used to represent the size of the bubble.
If the data source has 4 columns the last column is used to group each data point by the unique value assigned to the last column. For each unique value from the last column, a new series will be added to the bubble chart.